requirements.txt
Django>=5.2
python-decouple
python-telegram-bot>=20
psycopg2-binary
openai
edge-tts
deep-translator
gtts
Pillow
tzdata
pytest
pytest-django
pytest-cov
black
flake8
mypy
